fix type

+++ 2021/aoc2021day01/src/main.rs
@@ -5,7 +5,7 @@ fn main() {
         "{}",
         include_str!("./input.txt")
             .split('\n')
-            .filter_map(|s| s.parse::<usize>().ok())
+            .filter_map(|s| s.parse::<i32>().ok())
             .tuple_windows::<(_, _, _)>()
             .tuple_windows()
             .filter(|(a, b)| b.2 > a.0)
